Follow up on added ability to rewrite symbol references in BoundTreeRewriter.
- Extract common reusable functionality from MethodToClassRewriter into a new base class
- Remove some redundant overrides
- Adjust bound nodes and BoundTreeRewriter in preporation for extension method rewrite.
