+++ src/Compilers/CSharp/Portable/Binder/SwitchBinder_Patterns.cs
@@ -62,7 +62,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                 syntax: node,
                 expression: boundSwitchGoverningExpression,
                 innerLocals: locals,
-                innerLocalFunctions: functions,
+                innerLocalFunctions: ImmutableArray<MethodSymbol>.CastUp(functions),
                 switchSections: switchSections,
                 defaultLabel: defaultLabel,
                 breakLabel: this.BreakLabel,
