+++ src/Compilers/CSharp/Portable/Lowering/ClosureConversion/ClosureConversion.cs
@@ -765,7 +765,7 @@ namespace Microsoft.CodeAnalysis.CSharp
                         }
 
                         LocalSymbol localToUse;
-                        if (!localMap.TryGetValue(local, out localToUse))
+                        if (!TryGetRewrittenLocal(local, out localToUse))
                         {
                             localToUse = local;
                         }
@@ -1191,17 +1191,15 @@ namespace Microsoft.CodeAnalysis.CSharp
         public override BoundNode VisitScope(BoundScope node)
         {
             Debug.Assert(!node.Locals.IsEmpty);
-            var newLocals = ArrayBuilder<LocalSymbol>.GetInstance();
-            RewriteLocals(node.Locals, newLocals);
+            var newLocals = VisitLocals(node.Locals);
 
             var statements = VisitList(node.Statements);
-            if (newLocals.Count == 0)
+            if (newLocals.Length == 0)
             {
-                newLocals.Free();
                 return new BoundStatementList(node.Syntax, statements);
             }
 
-            return node.Update(newLocals.ToImmutableAndFree(), statements);
+            return node.Update(newLocals, statements);
         }
 
         public override BoundNode VisitCatchBlock(BoundCatchBlock node)
